+++ CHANGELOG.md
@@ -2,6 +2,26 @@
 
 All notable changes to this project will be documented in this file. See [standard-version](https://github.com/conventional-changelog/standard-version) for commit guidelines.
 
+## [8.4.0](https://github.com/GetStream/stream-chat-go/compare/v8.3.0...v8.4.0) (2026-02-25)
+
+
+### Features
+
+* remove filterTags from channel batch update API ([#376](https://github.com/GetStream/stream-chat-go/issues/376)) ([4bfd8cf](https://github.com/GetStream/stream-chat-go/commit/4bfd8cf574cefda8a4062b10c9fe8f5aad761fbb))
+
+## [8.3.0](https://github.com/GetStream/stream-chat-go/compare/v8.2.2...v8.3.0) (2026-02-24)
+
+### [8.2.2](https://github.com/GetStream/stream-chat-go/compare/v8.2.1...v8.2.2) (2026-02-19)
+
+
+### Features
+
+* add queryTeamUsageStats API endpoint ([#371](https://github.com/GetStream/stream-chat-go/issues/371)) ([d550df9](https://github.com/GetStream/stream-chat-go/commit/d550df995022ea4972b5aa95c043663a4a50e6ae))
+
+### [8.2.1](https://github.com/GetStream/stream-chat-go/compare/v8.2.0...v8.2.1) (2026-02-18)
+
+## [8.2.0](https://github.com/GetStream/stream-chat-go/compare/v8.1.1...v8.2.0) (2026-02-04)
+
 ### [8.1.1](https://github.com/GetStream/stream-chat-go/compare/v8.1.0...v8.1.1) (2026-01-08)